12/12

+++ ex01/iter.hpp
@@ -1,17 +1,15 @@
 #ifndef ITER_HPP_
 #define ITER_HPP_
 
-#include <cstddef>
-
 template <typename T, typename F>
-void iter(T* array, size_t len, const F& func) {
-    for (size_t i = 0; i < len; ++i)
+void iter(T* array, unsigned int len, const F& func) {
+    for (unsigned int i = 0; i < len; ++i)
         func(array[i]);
 }
 
 template <typename T, typename F>
-void iter(const T* array, size_t len, const F& func) {
-    for (size_t i = 0; i < len; ++i)
+void iter(const T* array, unsigned int len, const F& func) {
+    for (unsigned int i = 0; i < len; ++i)
         func(array[i]);
 }
